Pretty big refactor to support easy dockerizing of tests and server..
Most of the interesting work is in the added `docker-app` module.

Used `-` between modules names to be closer to standards.  I also
significantly reduced most of the package names.  We aren't producing
libraries;)
