+++ asynchttpclient-benchmark/src/main/java/benchmark/asynchttpclient/PerformanceTests.java
@@ -1,0 +1,18 @@
+package benchmark.asynchttpclient;
+
+import com.ss.benchmark.httpclient.common.BasePerformanceTest;
+import com.ss.benchmark.httpclient.common.HttpClientEngine;
+import org.testng.annotations.Test;
+
+/**
+ * @author sharath.srinivasa
+ */
+@Test(groups = "performance")
+public class PerformanceTests extends BasePerformanceTest {
+
+    @Override
+    protected HttpClientEngine getClient() {
+        return new Engine();
+    }
+}
+
