+++ rxnetty-benchmark/pom.xml
@@ -9,25 +9,57 @@
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>asynchttpclient-benchmark</artifactId>
+
+    <artifactId>rxnetty-benchmark</artifactId>
+
+    <properties>
+        <netty.version>4.1.5.Final</netty.version>
+        <rxnetty.version>0.5.2</rxnetty.version>
+    </properties>
 
     <dependencies>
-        <dependency>
-            <groupId>org.testng</groupId>
-            <artifactId>testng</artifactId>
-        </dependency>
         <dependency>
             <groupId>com.ss.benchmark</groupId>
             <artifactId>benchmark-common</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.asynchttpclient</groupId>
-            <artifactId>async-http-client</artifactId>
-            <version>2.6.0</version>
+            <groupId>org.testng</groupId>
+            <artifactId>testng</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>io.reactivex</groupId>
+            <artifactId>rxnetty-http</artifactId>
+            <version>${rxnetty.version}</version>
+        </dependency>
+        <!-- This was painful https://github.com/ReactiveX/RxNetty/issues/292 -->
+        <dependency>
+            <groupId>io.reactivex</groupId>
+            <artifactId>rxnetty-common</artifactId>
+            <version>${rxnetty.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.netty</groupId>
+            <artifactId>netty-buffer</artifactId>
+            <version>${netty.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.netty</groupId>
+            <artifactId>netty-codec-http</artifactId>
+            <version>${netty.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.netty</groupId>
+            <artifactId>netty-transport</artifactId>
+            <version>${netty.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.reactivex</groupId>
+            <artifactId>rxjava</artifactId>
+            <version>1.2.6</version>
         </dependency>
     </dependencies>
-    <build>
+        <build>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
